% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{docs}
\alias{docs}
\title{Open package documentation}
\usage{
docs()
}
\description{
This function launches a browser window for the \code{zkgr} package
documentation. The documentation is automatically rendered using
\href{https://pkgdown.r-lib.org/index.html}{pkgdown}.
}
\examples{
if(interactive()){
  docs() # open documentation in Web browser if using standard account
}
}
