% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{update_docs}
\alias{update_docs}
\title{Render pkgdown documentation for local access within package}
\usage{
update_docs()
}
\value{
\code{NULL} creates updated directory of pkgdown documentation
}
\description{
This function is intended to facilitate the maintenance and development of
this R package by copying pkgdown documentation from its location outside of
resources published with the package to a directory released with the package.
}
\examples{
if(FALSE){
  update_docs()
}
}
\keyword{internal}
