% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dbhds_docs.R
\name{use_dbhds_docs}
\alias{use_dbhds_docs}
\title{Use DBHDS documentation standards in your package}
\usage{
use_dbhds_docs(package = NULL)
}
\arguments{
\item{package}{The name of the package to create the documentation in (will
default to the current package name using the DESCRIPTION file in the active
.Rproj file)}
}
\description{
Does setup necessary to use self-contained DBHDS configurations for pkgdown
sites in your package. This includes the following functions:
\enumerate{
\item \code{update_docs()} - used to render a \code{pkgdown} static site within the
installed package, making the documentation available insite the package with
each release
\item \code{docs()} - when called from the package, will open a Web browser window
referencing the static HTML files for the \code{pkgdown} site rendered using
\code{update_docs()}
\item \code{serve_docs()} - when called from the package, will spin up a server on
the user's machine by calling \code{servr::httd()} to load the \code{pkgdown} site
stored within the package
}
}
\note{
This function requires the use roxygen.
}
\examples{
\dontrun{
use_dbhds_docs()
}

}
\seealso{
\code{vignette("documentation-functions", package = "zkgr")}
}
