% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dbhds_license.R
\name{use_dbhds_license}
\alias{use_dbhds_license}
\title{Use standard DBHDS package license}
\usage{
use_dbhds_license(copyright_holder)
}
\arguments{
\item{copyright_holder}{\code{character} The office within DBHDS that owns and
maintains the package}
}
\value{
A logical vector indicating if file was modified
}
\description{
Call this function to create a proprietary license aligned to DBHDS standards.
The license accepts an office name as the copyright holder.
}
\examples{
\dontrun{
  use_dbhds_license("Office of Epidemiology and Health Analytics")
}
}
