% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dbhds_logo.R
\name{use_dbhds_logo}
\alias{use_dbhds_logo}
\title{Use DBHDS logo in destination package}
\usage{
use_dbhds_logo(.overwrite = TRUE)
}
\arguments{
\item{.overwrite}{\code{logical}}
}
\value{
NULL copies the DBHDS logo to the destination package
}
\description{
Use this function to copy the DBHDS logo into the destination package. In
order to use this package with pkgdown, the DBHDS logo is saved in
'man/figures/logo.svg'.
}
\examples{
if(FALSE) use_dbhds_logo()
}
