% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dbhds_readme.R
\name{use_dbhds_readme}
\alias{use_dbhds_readme}
\title{Use DBHDS README.Rmd template}
\usage{
use_dbhds_readme(package = NULL)
}
\arguments{
\item{package}{The package name. If \code{null}, defaults to the current package
from the description file}
}
\value{
\code{null} The file README.Rmd will be created in the package
directory root
}
\description{
Creates an Rmarkdown file to be rendered as the package
README.md file. Importantly, the file must be knit as a standard markdown
file using \code{devtools::build_readme()}.
}
\details{
An R package is required to contain a README.md file. That file is a
standard markdown document with a \code{.md} extension. Using Rmarkdown,
it is possible to make a more engaging and visually appealing README file.

A basic README.Rmd file can be created within the package root directory by
calling \code{usethis::use_readme_rmd()}. The README.Rmd file created using \code{zkgr}
includes additional details and formatting aligned to DBHDS standards.
}
\examples{
\dontrun{
   use_dbhds_readme()
}
}
