% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pkgr.R
\name{use_pkgr}
\alias{use_pkgr}
\title{Use all default styles for DBHDS R packages}
\usage{
use_pkgr(
  copyright_holder,
  config_file = "_pkgdown.yml",
  destdir = "inst/pkgdown_docs",
  logo_overwrite = TRUE
)
}
\arguments{
\item{copyright_holder}{\code{character} the office who holds the proprietary
license within DBHDS}

\item{config_file}{\code{character} the YAML file used to customize the
pkgdown documentation site}

\item{destdir}{\code{character} the directory where pkgdown static files
will be rendered}

\item{logo_overwrite}{\code{logical} whether to overwrite an existing logo file}
}
\description{
\code{use_pkgr} is a wrapper function for all functions within
the \code{zkgr} package. It allows the user to rapidly adopt the default
DBHDS package license, pkgdown template, and agency logo.

The command

\if{html}{\out{<div class="sourceCode R">}}\preformatted{use_pkgr("Office of Epidemiology and Health Analytics")
}\if{html}{\out{</div>}}

is equivalent to calling the following functions in order:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{usethis::use_roxygen_md()   # enable markdown in package .Rd files
use_dbhds_license("Office of Epidemiology and Health Analytics")
devtools::document()        # build package metadata for first time
use_dbhds_docs()
use_pkgr_yaml()
use_dbhds_logo()
devtools::document()
use_dbhds_readme()
devtools::build_readme()
}\if{html}{\out{</div>}}

Most users of this package will find this single function to be the only one
they need to use for this package.
}
\examples{
\dontrun{
   # create files in current package
   use_pkgr(copyright_holder = "Office of Epidemiology and Health Analytics")
}
}
