% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pkgr_yaml.R
\name{use_pkgr_yaml}
\alias{use_pkgr_yaml}
\title{Use pkgr YAML file to load styles from this package}
\usage{
use_pkgr_yaml(
  config_file = "_pkgdown.yml",
  destdir = "inst/pkgdown_docs",
  strict = TRUE
)
}
\arguments{
\item{config_file}{Path to the pkgdown yaml config file, relative to the
project}

\item{destdir}{Target directory for pkgdown docs}

\item{strict}{\code{logical} If \code{TRUE}, requires that prompts be answered in an
interactive session}
}
\description{
\href{https://pkgdown.r-lib.org}{pkgdown} makes it easy to turn your package into
a beautiful website.

\code{use_pkgr_yaml} creates a YAML file  entitled \verb{_pkgdown.yml} within a package
that automatically includes a reference to this package's template styles.
The structure of the YAML file is very basic, but the goal of this function
is to facilitate rapid development and minimize the amount of boilerplate
code necessary to style a new package.

Below is the structure of the YAML file.

\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{template:
   package: zkgr
}\if{html}{\out{</div>}}

Using this YAML template will enable a user to adopt styles aligned to DBHDS
package guidelines with a single line of code.
}
\seealso{
\url{https://pkgdown.r-lib.org/articles/customise.html#template-packages}
}
